Restructuring & documentation
